+++ core/domain/src/main/java/com/wansenai/bo/sale/SalesDataBO.java
@@ -10,9 +10,11 @@
  * OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions
  * and limitations under the License.
  */
-package com.wansenai.bo;
+package com.wansenai.bo.sale;
 
-import com.wansenai.utils.excel.ExcelExport;
+import com.fasterxml.jackson.annotation.JsonFormat;
+import com.fasterxml.jackson.databind.annotation.JsonSerialize;
+import com.wansenai.bo.BigDecimalSerializerBO;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -24,56 +26,46 @@ import java.math.BigDecimal;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class PurchaseDataExportBO {
+public class SalesDataBO {
 
-    @ExcelExport(value = "供应商")
-    private String supplierName;
+    @JsonFormat(shape = JsonFormat.Shape.STRING)
+    private Long warehouseId;
 
-    @ExcelExport(value = "单据编号")
-    private String receiptNumber;
-
-    @ExcelExport(value = "仓库名称")
     private String warehouseName;
 
-    @ExcelExport(value = "条码")
     private String barCode;
 
-    @ExcelExport(value = "商品名称")
+    @JsonFormat(shape = JsonFormat.Shape.STRING)
+    private Long productId;
+
     private String productName;
 
-    @ExcelExport(value = "规格")
     private String productStandard;
 
-    @ExcelExport(value = "型号")
     private String productModel;
 
-    @ExcelExport(value = "颜色")
     private String productColor;
 
-    @ExcelExport(value = "库存")
     private Integer stock;
 
-    @ExcelExport(value = "单位")
     private String productUnit;
 
-    @ExcelExport(value = "数量")
     private Integer productNumber;
 
-    @ExcelExport(value = "单价")
+    @JsonSerialize(using = BigDecimalSerializerBO.class)
     private BigDecimal unitPrice;
 
-    @ExcelExport(value = "金额")
+    @JsonSerialize(using = BigDecimalSerializerBO.class)
     private BigDecimal amount;
 
-    @ExcelExport(value = "税率(%)")
+    @JsonSerialize(using = BigDecimalSerializerBO.class)
     private BigDecimal taxRate;
 
-    @ExcelExport(value = "税额")
+    @JsonSerialize(using = BigDecimalSerializerBO.class)
     private BigDecimal taxAmount;
 
-    @ExcelExport(value = "价税合计")
+    @JsonSerialize(using = BigDecimalSerializerBO.class)
     private BigDecimal taxTotalPrice;
 
-    @ExcelExport(value = "备注")
     private String remark;
 }
